api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
asyncpg>=0.30.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
google-generativeai>=0.8.0
openai>=1.40.0
anthropic>=0.34.0
python-multipart>=0.0.9
gunicorn>=23.0.0
psycopg2-binary>=2.9.9
PyPDF2>=3.0.0